+++ samples/Get-TopProcess.ps1
@@ -13,12 +13,12 @@ Function Get-TopProcess {
         [string]$Computername = $env:computername
     )
     Begin {
-        $vlog = New-CustomFileName -Template "$env:temp\$($MyInvocation.MyCommand)-%time.log"
-        $PSDefaultParameterValues."Out-VerboseTee:Path" = $vlog
+        $vLog = New-CustomFileName -Template "$env:temp\$($MyInvocation.MyCommand)-%time.log"
+        $PSDefaultParameterValues."Out-VerboseTee:Path" = $vLog
         $PSDefaultParameterValues."Out-VerboseTee:Append" = $True
         $PSDefaultParameterValues."Write-Detail:Prefix" = 'Begin'
         Write-Detail "Starting $($MyInvocation.MyCommand)" | Tee-Verbose
-        Write-Detail "Using verbose log $vlog" | Tee-Verbose
+        Write-Detail "Using verbose log $vLog" | Tee-Verbose
         Write-Detail "Execution metadata" | Tee-Verbose
         Write-Detail (Add-Border -TextBlock (Get-PSWho -AsString) | Out-String) | Tee-Verbose
         Write-Detail "Initializing data array" | Tee-Verbose
@@ -40,8 +40,8 @@ Function Get-TopProcess {
         if ($VerbosePreference -eq 'Continue') {
             #save results to the verbose log
             #add the detail to the log but don't make it verbose
-            Write-Detail "Data results" | Add-Content -Path $vlog
-            $all | Out-String | Add-Content -Path $vlog
+            Write-Detail "Data results" | Add-Content -Path $vLog
+            $all | Out-String | Add-Content -Path $vLog
             #save results for each computer to a separate log
             $all | Group-Object -Property PSComputername |
                 ForEach-Object {