+++ samples/ProcessPercent.ps1
@@ -6,7 +6,8 @@ $total = ($processes | Measure-Object -Property WS -Sum).sum
 
 $processes | Sort-Object WS -Descending |
     Select-Object -First 25 |
-    Format-Table -Property Name, ID, Handles, @{Name = 'WS(MB)'; Expression = { Format-Value $_.WS -Unit MB } },
+    Format-Table -Property Name, ID, Handles,
+    @{Name = 'WS(MB)'; Expression = { Format-Value $_.WS -Unit MB } },
     @{Name = "PctWS"; Expression = {
             $pct = Format-Percent -Value $_.ws -Total $total -Decimal 2
             $bar = New-ANSIBar -Range 14 -Spacing $pct -Character BlackSquare